wip: replace hardcoded request values with dynamic consumption from the form schema
