+++ src/NextGen/DonationForm/DataTransferObjects/DonateControllerData.php
@@ -1,0 +1,110 @@
+<?php
+
+namespace Give\NextGen\DonationForm\DataTransferObjects;
+
+use Give\Donations\Models\Donation;
+use Give\Donations\ValueObjects\DonationStatus;
+use Give\Framework\Support\ValueObjects\Money;
+use Give\NextGen\DonationForm\Models\DonationForm;
+
+class DonateControllerData
+{
+    /**
+     * @var float
+     */
+    public $amount;
+    /**
+     * @var string
+     */
+    public $gatewayId;
+    /**
+     * @var string
+     */
+    public $currency;
+    /**
+     * @var string
+     */
+    public $firstName;
+    /**
+     * @var string
+     */
+    public $lastName;
+    /**
+     * @var string
+     */
+    public $email;
+    /**
+     * @var int
+     */
+    public $wpUserId;
+    /**
+     * @var int
+     */
+    public $formId;
+    /**
+     * @var string
+     */
+    public $formTitle;
+    /**
+     * @var string|null
+     */
+    public $company;
+    /**
+     * @var string|null
+     */
+    public $honorific;
+
+      /**
+     * @unreleased
+     */
+    public function toDonation($donorId): Donation
+    {
+        return new Donation([
+            'status' => DonationStatus::PENDING(),
+            'gatewayId' => $this->gatewayId,
+            'amount' => Money::fromDecimal($this->amount, $this->currency),
+            'donorId' => $donorId,
+            'firstName' => $this->firstName,
+            'lastName' => $this->lastName,
+            'email' => $this->email,
+            'formId' => $this->formId,
+            'formTitle' => get_the_title($this->formId),
+            'company' => $this->company
+        ]);
+    }
+
+     /**
+     * @unreleased
+     */
+    public function getDonationForm(): DonationForm
+    {
+        return DonationForm::find($this->formId);
+    }
+
+    /**
+     * @unreleased
+     */
+    public function getCustomFields(): array
+    {
+        $properties = get_object_vars($this);
+
+         return array_filter($properties, static function ($param) {
+            return !in_array(
+                $param,
+                [
+                    'amount',
+                    'gatewayId',
+                    'currency',
+                    'firstName',
+                    'lastName',
+                    'email',
+                    'wpUserId',
+                    'formId',
+                    'formTitle',
+                    'company',
+                    'honorific',
+                ]
+            );
+        }, ARRAY_FILTER_USE_KEY);
+    }
+}
